Refactor anchor auto-stitch workflow
- Rename stitch code into a clearer package structure:
  stitch_tasks.py plus stitch/runner.py, stitch/core/, and stitch/configs/
- Replace CLI-style stitch invocation with direct Python imports and function calls
- Simplify the Prefect entrypoint so it prepares run metadata and calls run_stitch_validation()
- Remove redundant path-loader layers, compatibility wrappers, debug helpers, and unused functions
- Route stitch outputs through the same proposal/experiment directory logic used by linker.py
- Update tests and docs for the new module names and output path behavior
