+++ pixi.toml
@@ -16,5 +16,4 @@ event-model = "*"
 nslsii = "*"
 python-dotenv = ">=1.2.2,<2"
 numpy = ">=1.26"
-matplotlib = ">=3.8"
 pillow = ">=10"
