+++ stitch/README.md
@@ -1,13 +1,12 @@
-# Bundled Auto-Stitch Module
+# Stitch Package
 
 This folder contains the minimum Part 2 stitching runtime copied into `cms-workflows` so the Prefect flow can run standalone from this repository.
 
 Included components:
 
-- `stitch.py`
-- `examples/phase1_scan_range_from_tiled.py`
-- `stitching/` package
+- `runner.py`
+- `core/` package
 - `configs/stitching_defaults.json`
 - `configs/Dectris/` masks required by the default config
 
-`end_of_run_workflow.py` uses this bundled folder by default when `stitch_repo_path` is not set.
+`stitch_tasks.py` imports `run_stitch_validation` from `stitch.runner` directly.
