+++ stitch/__init__.py
@@ -1,0 +1,1 @@
+"""Stitch package used by the Prefect stitch tasks."""
