+++ stitch/configs/Dectris/Eiger1M_main_gaps-mask.png
