+++ stitch/configs/Dectris/Eiger4M_all_gaps-mask.png
