+++ stitch/configs/Dectris/Eiger4M_main_gaps-mask.png
