+++ stitch/configs/Dectris/Pilatus1M_generic-mask.png
