+++ stitch/configs/Dectris/Pilatus1M_main_gaps-mask.png
