+++ stitch/configs/Dectris/Pilatus2M_gaps-mask.png
