+++ stitch/configs/Dectris/Pilatus2M_generic-mask.png
