+++ stitch/configs/Dectris/Pilatus2M_vertical_gaps-mask.png
