+++ stitch/configs/Dectris/Pilatus300kWh_generic-mask.png
