+++ stitch/configs/Dectris/Pilatus300kWh_main_gaps-mask.png
