+++ stitch/configs/Dectris/Pilatus300kWv_generic-mask.png
