+++ stitch/configs/Dectris/Pilatus300kWv_main_gaps-mask.png
