+++ stitch/configs/Dectris/Pilatus300k_generic-mask.png
