+++ stitch/configs/Dectris/Pilatus300k_main_gaps-mask.png
