+++ stitch/configs/Dectris/Pilatus800k2_corner-mask.png
