+++ stitch/configs/Dectris/Pilatus800k2_gaps-mask.png
