+++ stitch/configs/Dectris/Pilatus800k2_generic-mask.png
