+++ stitch/configs/Dectris/Pilatus800k2_vertical_gaps-mask.png
