+++ stitch/configs/Dectris/Pilatus800k_corner-mask.png
