+++ stitch/configs/Dectris/Pilatus800k_gaps-mask.png
