+++ stitch/configs/Dectris/Pilatus800k_generic-mask.png
