+++ stitch/configs/Dectris/Pilatus800k_vertical_gaps-mask.png
