+++ stitch/configs/stitching_defaults.json
