+++ stitch/core/__init__.py
