+++ stitch/core/adapters/__init__.py
