+++ stitch/core/adapters/tiled_adapter.py
