+++ stitch/core/config.py
