+++ stitch/core/core.py
