+++ stitch/core/errors.py
