+++ stitch/core/grouping.py
