+++ stitch/core/mask.py
