+++ stitch/core/models.py
