+++ stitch/core/modes.py
