+++ stitch/core/serialize.py
