+++ test/test_linker_paths.py
@@ -47,6 +47,18 @@ class MakeRelativePathTests(unittest.TestCase):
             proposal / "experiments",
         )
 
+    def test_experiment_alias_directory_uses_proposal_layout(self):
+        doc = {
+            "cycle": "2026-1",
+            "data_session": "pass-12345",
+            "experiments_directory": "experiments",
+            "experiment_alias_directory": "sample-a",
+        }
+        self.assertEqual(
+            linker.experiment_alias_directory(doc),
+            Path("/nsls2/data/cms/proposals/2026-1/pass-12345/experiments/sample-a"),
+        )
+
 
 if __name__ == "__main__":
     unittest.main()
